# docs/swagger.yaml
basePath: /
definitions:
  handler.ShortenURLRequest:
    properties:
      exp:
        example: 3600
        maximum: 3600
        type: integer
      url:
        example: https://example.com
        type: string
    required:
    - exp
    - url
    type: object
  handler.ShortenURLResponse:
    properties:
      code:
        example: abc123
        type: string
      message:
        example: Shorten URL generated successfully!
        type: string
    type: object
  handler.healthCheckResponse:
    properties:
      error:
        type: string
      instance_id:
        example: instance-test
        type: string
      message:
        example: OK
        type: string
      service_name:
        example: bookmark_service
        type: string
    type: object
  handler.loginInputBody:
    properties:
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.loginResBody:
    properties:
      data:
        type: string
      message:
        type: string
    type: object
  handler.profileResBody:
    properties:
      data:
        $ref: '#/definitions/model.User'
    type: object
  handler.registerInputBody:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  handler.registerResBody:
    properties:
      data:
        $ref: '#/definitions/handler.registerUserData'
      message:
        type: string
    type: object
  handler.registerUserData:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  handler.updateProfileInputBody:
    properties:
      display_name:
        type: string
      email:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  response.Response:
    properties:
      details: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for bookmark service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BookMark_Service API
  version: "1.2"
paths:
  /v1/gen-pass:
    get:
      description: Generates a cryptographically secure random password
      produces:
      - text/plain
      responses:
        "200":
          description: Generated password
          schema:
            type: string
        "500":
          description: Error message
          schema:
            type: string
      summary: Generate a random password
      tags:
      - password
  /v1/health-check:
    get:
      description: Health check
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.healthCheckResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Health check
      tags:
      - health_check
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      description: Get URL by code
      parameters:
      - description: Url code
        format: string
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request - invalid URL or validation error
        "404":
          description: URL not found
        "500":
          description: Internal Server Error
      summary: Get URL
      tags:
      - URL Shortener
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Shortens a given URL and returns a shortened URL code.
      parameters:
      - description: URL to shorten
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ShortenURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShortenURLResponse'
        "400":
          description: Bad Request - invalid URL or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Shorten URL
      tags:
      - URL Shortener
  /v1/self/info:
    get:
      consumes:
      - application/json
      description: Get profile of the currently authenticated user using JWT in Authorization
        header
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.profileResBody'
        "401":
          description: Invalid or missing token
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update display name and/or email of the currently authenticated
        user
      parameters:
      - description: Update profile request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateProfileInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: No data provided for update
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Invalid or missing token
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User does not exist
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with username and password, returns a JWT token
      parameters:
      - description: User login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.loginInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.loginResBody'
        "400":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login a user
      tags:
      - Users
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account using username, password, display name,
        and email.
      parameters:
      - description: User registration payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.registerInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.registerResBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register a new user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"